# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Credd AI - Financial infrastructure for autonomous AI agents
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentBody"
      responses:
        "201":
          description: Created agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
  /v1/agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
  /v1/agents/{agentId}/accounts:
    get:
      operationId: listAgentAccounts
      tags: [accounts]
      summary: List accounts for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
...
  /v1/accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /v1/accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /v1/accounts/{id}/status:
    patch:
      operationId: updateAccountStatus
      tags: [accounts]
      summary: Update account status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
...
  /v1/transactions/{id}/approval:
    get:
      operationId: getTransactionApproval
      tags: [approvals]
      summary: Get approval for a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
  /v1/approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      summary: List all approval requests
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of approval requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalListResponse"
  /v1/approvals/{id}:
    get:
      operationId: getApproval
      tags: [approvals]
      summary: Get approval by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approval request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalResponse"
  /v1/approvals/{id}/approve:
    post:
      operationId: approveRequest
      tags: [approvals]
      summary: Approve an approval request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveDeclineBody"
      responses:
        "200":
          description: Approval result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
  /v1/approvals/{id}/decline:
    post:
      operationId: declineRequest
      tags: [approvals]
      summary: Decline an approval request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveDeclineBody"
      responses:
        "200":
          description: Decline result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalDecisionResponse"
  /v1/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [auditLogs]
      summary: List audit logs
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
        - name: entityId
          in: query
          required: false
          schema:
            type: string
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
  /v1/audit-logs/{id}:
    get:
      operationId: getAuditLog
      tags: [auditLogs]
      summary: Get audit log by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
  /v1/entities/{entityType}/{entityId}/audit-logs:
    get:
      operationId: listEntityAuditLogs
      tags: [auditLogs]
      summary: List audit logs for an entity
      parameters:
        - name: entityType
          in: path
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
...
components:
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended, pending]
        description:
          type: string
          nullable: true
        externalRef:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt
    Account:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        name:
          type: string
        accountRef:
          type: string
          nullable: true
        currency:
          type: string
        status:
          type: string
          enum: [active, frozen, closed]
        balanceMinor:
          type: integer
        availableBalanceMinor:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agentId
        - name
        - currency
        - status
        - balanceMinor
        - availableBalanceMinor
        - createdAt
        - updatedAt
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        accountId:
          type: string
        cardId:
          type: string
          nullable: true
        merchantId:
          type: string
          nullable: true
        direction:
          type: string
          enum: [DEBIT, CREDIT]
        amountMinor:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [attempted, pending, approved, declined, settled, reversed]
        decision:
          type: string
          enum: [ALLOW, DENY, REVIEW, NONE]
        approvalStatus:
          type: string
          enum: [NOT_REQUIRED, PENDING, APPROVED, DECLINED]
        declinedReason:
          type: string
          nullable: true
        requestedAt:
          type: string
        processedAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - agentId
        - accountId
        - direction
        - amountMinor
        - currency
        - status
        - decision
        - approvalStatus
        - requestedAt
        - createdAt
        - updatedAt
    ApprovalRequest:
      type: object
      properties:
        id:
          type: string
        transactionId:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, DECLINED, EXPIRED]
        requestedByType:
          type: string
        requestedReason:
          type: string
          nullable: true
        reviewerId:
          type: string
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        decisionReason:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - transactionId
        - status
        - requestedByType
        - createdAt
        - updatedAt
    AuditLog:
      type: object
      properties:
        id:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        action:
          type: string
        actorType:
          type: string
        actorId:
          type: string
          nullable: true
        traceId:
          type: string
          nullable: true
        payloadJson:
          type: object
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - entityType
        - entityId
        - action
        - actorType
        - createdAt
...
    CreateAgentBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended, pending]
        description:
          type: string
      required:
        - name
    CreateAccountBody:
      type: object
      properties:
        agentId:
          type: string
        name:
          type: string
        currency:
          type: string
        accountRef:
          type: string
      required:
        - agentId
        - name
...
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    ApproveDeclineBody:
      type: object
      properties:
        reviewerId:
          type: string
        decisionReason:
          type: string

    # Response wrappers
    AgentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Agent"
      required:
        - success
        - data
    AgentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
      required:
        - success
        - data
    AccountResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Account"
      required:
        - success
        - data
    AccountListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Account"
      required:
        - success
        - data
...
    ApprovalResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/ApprovalRequest"
      required:
        - success
        - data
    ApprovalListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequest"
      required:
        - success
        - data
    ApprovalDecisionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            approval:
              $ref: "#/components/schemas/ApprovalRequest"
            transaction:
              $ref: "#/components/schemas/Transaction"
          required:
            - approval
            - transaction
      required:
        - success
        - data
    AuditLogResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AuditLog"
      required:
        - success
        - data
    AuditLogListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
      required:
        - success
        - data
...