# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Credd AI - Financial infrastructure for autonomous AI agents
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentBodyStatus } from "./createAgentBodyStatus";

export interface CreateAgentBody {
  name: string;
  status?: CreateAgentBodyStatus;
  description?: string;
}